requirements-file.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
schedule==1.2.0
python-dotenv==1.0.0
flask==2.3.3
fake-useragent==1.2.1
gunicorn==21.2.0
